requirements.txt
# ============================================================
# Install torch first, then install this file.
#
# CPU:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
#
# CUDA 12.1:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#
# Then:
# pip install -r requirements.txt
# ============================================================

# OCR backends
surya-ocr==0.6.0
easyocr==1.7.2

# Surya dependencies
transformers==4.44.2
huggingface-hub==0.23.4
tokenizers==0.19.1

# Document IO and utilities
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
tqdm==4.66.4
python-dotenv==1.0.1
pdfplumber==0.11.9
pypdfium2==4.30.0
python-docx==1.2.0
